+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.38.15] - 2025-09-22
+
+### Changed
+
+- Prevent pre-commit configuration during initialization and ensure -s option skips hooks
+- Prevent pre-commit configuration during initialization and ensure -s option skips hooks
+
+### Testing
+
+- test: Update 12 files
+
 ## [0.38.14] - 2025-09-22
 
 ### Testing